requirements.txt
PyJWT==1.7.1
boto3==1.12.39
botocore==1.15.39
Django==3.0.5
django-cors-headers==3.3.0
django-filter==2.2.0
django-jsonfield==1.4.0
django-storages==1.9.1
djangorestframework==3.11.0
mysqlclient==1.4.6
Pillow==7.1.2
requests==2.23.0
sendgrid==6.3.1